Move Half to headeronly
ghstack-source-id: 01e7cdaa64d10ab53121c2ff1690864296d0116e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/159172
